move files to appropriate dir
