+++ narwhal/config/src/lib.rs
@@ -135,6 +135,14 @@ pub struct Parameters {
     pub max_concurrent_requests: usize,
     /// Properties for the prometheus metrics
     pub prometheus_metrics: PrometheusMetricsParameters,
+    /// Network admin server port number
+    #[serde(default = "default_network_admin_server_port")]
+    pub network_admin_server_port: u16,
+}
+
+pub fn default_network_admin_server_port() -> u16 {
+    let host = "127.0.0.1";
+    get_available_port(host)
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -270,6 +278,7 @@ impl Default for Parameters {
             consensus_api_grpc: ConsensusAPIGrpcParameters::default(),
             max_concurrent_requests: 500_000,
             prometheus_metrics: PrometheusMetricsParameters::default(),
+            network_admin_server_port: default_network_admin_server_port(),
         }
     }
 }
@@ -341,6 +350,10 @@ impl Parameters {
         info!(
             "Prometheus metrics server will run on {}",
             self.prometheus_metrics.socket_addr
+        );
+        info!(
+            "Network admin server will run on 127.0.0.1:{}",
+            self.network_admin_server_port
         );
     }
 }
@@ -736,5 +749,6 @@ mod tests {
         assert!(logs_contain(
             "Prometheus metrics server will run on /ip4/127.0.0.1/tcp"
         ));
+        assert!(logs_contain("Network admin server will run on 127.0.0.1:"));
     }
 }
