+++ narwhal/network/Cargo.toml
@@ -24,12 +24,13 @@ types = { path = "../types", package = "narwhal-types" }
 crypto = { path = "../crypto", package = "narwhal-crypto" }
 
 serde = "1.0.144"
-workspace-hack.workspace = true
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
 eyre = "0.6.8"
 
 anemo = { git = "https://github.com/mystenlabs/anemo.git", rev = "302ff6f98fffa30a3a18e919766cd63f5b4c5aa8" }
 anemo-tower = { git = "https://github.com/mystenlabs/anemo.git", rev = "302ff6f98fffa30a3a18e919766cd63f5b4c5aa8" }
 anyhow = "1.0.65"
+axum = "0.5.16"
 
 [dev-dependencies]
 bincode = "1.3.3"
