+++ narwhal/network/src/metrics.rs
@@ -1,14 +1,38 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
-use anemo_tower::callback::MakeCallbackHandler;
-use anemo_tower::callback::ResponseHandler;
-use prometheus::HistogramTimer;
+use anemo_tower::callback::{MakeCallbackHandler, ResponseHandler};
 use prometheus::{
-    register_histogram_vec_with_registry, register_int_counter_vec_with_registry,
-    register_int_gauge_vec_with_registry, IntGaugeVec, Registry,
+    default_registry, register_histogram_vec_with_registry, register_int_counter_vec_with_registry,
+    register_int_gauge_vec_with_registry, HistogramTimer, HistogramVec, IntCounterVec, IntGaugeVec,
+    Registry,
 };
-use prometheus::{HistogramVec, IntCounterVec};
 use std::sync::Arc;
+
+#[derive(Clone, Debug)]
+pub struct NetworkConnectionMetrics {
+    /// The connection status of a peer. 0 if not connected, 1 if connected.
+    pub network_peer_connected: IntGaugeVec,
+}
+
+impl NetworkConnectionMetrics {
+    pub fn new(registry: &Registry) -> Self {
+        Self {
+            network_peer_connected: register_int_gauge_vec_with_registry!(
+                "network_peer_connected",
+                "The connection status of a peer. 0 if not connected, 1 if connected",
+                &["peer_id"],
+                registry
+            )
+            .unwrap(),
+        }
+    }
+}
+
+impl Default for NetworkConnectionMetrics {
+    fn default() -> Self {
+        Self::new(default_registry())
+    }
+}
 
 #[derive(Clone)]
 pub struct NetworkMetrics {
