+++ narwhal/primary/Cargo.toml
@@ -41,9 +41,10 @@ crypto = { path = "../crypto", package = "narwhal-crypto" }
 network = { path = "../network", package = "narwhal-network" }
 types = { path = "../types", package = "narwhal-types" }
 storage = { path = "../storage", package = "narwhal-storage" }
-mysten-network.workspace = true
-store.workspace = true
-workspace-hack.workspace = true
+mysten-network = "0.1.0"
+
+store = { version = "0.1.0", package = "typed-store" }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 anemo = { git = "https://github.com/mystenlabs/anemo.git", rev = "302ff6f98fffa30a3a18e919766cd63f5b4c5aa8" }
 anemo-tower = { git = "https://github.com/mystenlabs/anemo.git", rev = "302ff6f98fffa30a3a18e919766cd63f5b4c5aa8" }
@@ -58,13 +59,14 @@ mockall = "0.11.2"
 node = { path = "../node", package = "narwhal-node" }
 proptest = "1.0.0"
 tempfile = "3.3.0"
-telemetry-subscribers.workspace = true
+telemetry-subscribers = "0.1.0"
 test-utils = { path = "../test-utils", package = "narwhal-test-utils" }
 thiserror = "1.0.35"
 tracing = "0.1.36"
 tracing-test = "0.2.3"
 worker = { path = "../worker", package = "narwhal-worker" }
 storage = { path = "../storage", package = "narwhal-storage" }
+reqwest = { version = "0.11.12", features = ["json"] }
 
 [features]
 benchmark = []
