+++ narwhal/primary/src/metrics.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 use crate::EndpointMetrics;
 use mysten_network::metrics::MetricsCallbackProvider;
-use network::metrics::NetworkMetrics;
+use network::metrics::{NetworkConnectionMetrics, NetworkMetrics};
 use prometheus::{
     core::{AtomicI64, GenericGauge},
     default_registry, register_histogram_vec_with_registry, register_int_counter_vec_with_registry,
@@ -20,6 +20,7 @@ pub(crate) struct Metrics {
     pub(crate) outbound_network_metrics: Option<NetworkMetrics>,
     pub(crate) primary_channel_metrics: Option<PrimaryChannelMetrics>,
     pub(crate) node_metrics: Option<PrimaryMetrics>,
+    pub(crate) network_connection_metrics: Option<NetworkConnectionMetrics>,
 }
 
 /// Initialises the metrics
@@ -37,12 +38,16 @@ pub(crate) fn initialise_metrics(metrics_registry: &Registry) -> Metrics {
     // Essential/core metrics across the primary node
     let node_metrics = PrimaryMetrics::new(metrics_registry);
 
+    // Network metrics for the primary connection
+    let network_connection_metrics = NetworkConnectionMetrics::new(metrics_registry);
+
     Metrics {
         node_metrics: Some(node_metrics),
         endpoint_metrics: Some(endpoint_metrics),
         primary_channel_metrics: Some(primary_channel_metrics),
         inbound_network_metrics: Some(inbound_network_metrics),
         outbound_network_metrics: Some(outbound_network_metrics),
+        network_connection_metrics: Some(network_connection_metrics),
     }
 }
