+++ narwhal/worker/Cargo.toml
@@ -30,15 +30,15 @@ crypto = { path = "../crypto", package = "narwhal-crypto" }
 network = { path = "../network", package = "narwhal-network" }
 primary = { path = "../primary", package = "narwhal-primary" }
 types = { path = "../types", package = "narwhal-types" }
-mysten-network.workspace = true
+mysten-network = "0.1.0"
 prometheus = "0.13.2"
 
 anemo = { git = "https://github.com/mystenlabs/anemo.git", rev = "302ff6f98fffa30a3a18e919766cd63f5b4c5aa8" }
 anemo-tower = { git = "https://github.com/mystenlabs/anemo.git", rev = "302ff6f98fffa30a3a18e919766cd63f5b4c5aa8" }
 anyhow = "1.0.65"
 
-store.workspace = true
-workspace-hack.workspace = true
+store = { version = "0.1.0", package = "typed-store" }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 [dev-dependencies]
 arc-swap = { version = "1.5.1", features = ["serde"] }
@@ -46,6 +46,10 @@ rand = "0.8.5"
 telemetry-subscribers.workspace = true
 tempfile = "3.3.0"
 test-utils = { path = "../test-utils", package = "narwhal-test-utils" }
+reqwest = { version = "0.11.12", features = ["json"] }
+node = { path = "../node", package = "narwhal-node" }
+consensus = { path = "../consensus", package = "narwhal-consensus" }
+primary = { path = "../primary", package = "narwhal-primary" }
 
 [features]
 benchmark = []
