+++ narwhal/worker/src/metrics.rs
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 use mysten_network::metrics::MetricsCallbackProvider;
-use network::metrics::NetworkMetrics;
+use network::metrics::{NetworkConnectionMetrics, NetworkMetrics};
 use prometheus::{
     default_registry, register_histogram_vec_with_registry, register_int_counter_vec_with_registry,
     register_int_counter_with_registry, register_int_gauge_vec_with_registry,
@@ -18,6 +18,7 @@ pub struct Metrics {
     pub endpoint_metrics: Option<WorkerEndpointMetrics>,
     pub inbound_network_metrics: Option<NetworkMetrics>,
     pub outbound_network_metrics: Option<NetworkMetrics>,
+    pub network_connection_metrics: Option<NetworkConnectionMetrics>,
 }
 
 /// Initialises the metrics
@@ -35,12 +36,16 @@ pub fn initialise_metrics(metrics_registry: &Registry) -> Metrics {
     let inbound_network_metrics = NetworkMetrics::new("worker", "inbound", metrics_registry);
     let outbound_network_metrics = NetworkMetrics::new("worker", "outbound", metrics_registry);
 
+    // Network metrics for the worker connection
+    let network_connection_metrics = NetworkConnectionMetrics::new(metrics_registry);
+
     Metrics {
         worker_metrics: Some(node_metrics),
         channel_metrics: Some(channel_metrics),
         endpoint_metrics: Some(endpoint_metrics),
         inbound_network_metrics: Some(inbound_network_metrics),
         outbound_network_metrics: Some(outbound_network_metrics),
+        network_connection_metrics: Some(network_connection_metrics),
     }
 }
